Method to get the default table to use.

It can accept a list of new padj values matching the same dimmensions than the current vector. It can calculate the lfdr based on fdrtool::fdrtool function.

degDefault(object)

degCorrect(object, fdr)

# S4 method for DEGSet
degDefault(object)

# S4 method for DEGSet
degCorrect(object, fdr)

Arguments

object
DEGSet

fdr
It can be fdr-stat, fdr-pvalue, vector of new padj

object
DEGSet

Examples

library(DESeq2)
library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following object is masked from ‘package:matrixStats’:
#> 
#>     count

#> The following object is masked from ‘package:Biobase’:
#> 
#>     combine

#> The following objects are masked from ‘package:GenomicRanges’:
#> 
#>     intersect, setdiff, union

#> The following object is masked from ‘package:GenomeInfoDb’:
#> 
#>     intersect

#> The following objects are masked from ‘package:IRanges’:
#> 
#>     collapse, desc, intersect, setdiff, slice, union

#> The following objects are masked from ‘package:S4Vectors’:
#> 
#>     first, intersect, rename, setdiff, setequal, union

#> The following objects are masked from ‘package:BiocGenerics’:
#> 
#>     combine, intersect, setdiff, union

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

dds <- makeExampleDESeqDataSet(betaSD=1)
colData(dds)[["treatment"]] <- sample(colData(dds)[["condition"]], 12)
design(dds) <-  ~ condition + treatment
dds <- DESeq(dds)

#> estimating size factors

#> estimating dispersions

#> gene-wise dispersion estimates

#> mean-dispersion relationship

#> final dispersion estimates

#> fitting model and testing

res <- degComps(dds, contrast = list("treatment_B_vs_A"))

#> Doing 1 element(s).

#> Doing results() for each element.

#> Doing lcfSrink() for each element.

#> using 'normal' for LFC shrinkage, the Normal prior from Love et al (2014).
#> 
#> Note that type='apeglm' and type='ashr' have shown to have less bias than type='normal'.
#> See ?lfcShrink for more details on shrinkage type, and the DESeq2 vignette.
#> Reference: https://doi.org/10.1093/bioinformatics/bty895

degCorrect(res, fdr = "lfdr-stat")

#> Step 1... determine cutoff point
#> Step 2... estimate parameters of null distribution and eta0
#> Step 3... compute p-values and estimate empirical PDF/CDF
#> Step 4... compute q-values and local fdr
#> 
#> Step 1... determine cutoff point
#> Step 2... estimate parameters of null distribution and eta0
#> Step 3... compute p-values and estimate empirical PDF/CDF
#> Step 4... compute q-values and local fdr
#> 

#> Comparisons: treatment.B.vs.A
#> Results in comparison: raw,shrunken
#> Default is: shrunken