Filter genes by group

This function will keep only rows that have a minimum counts of 1 at least in a min number of samples (default 80

degFilter(counts, metadata, group, min = 0.8, minreads = 0)

Arguments

counts
Matrix with expression data, columns are samples and rows are genes or other feature.

metadata
Data.frame with information about each column in counts matrix. Rownames should match colnames(counts).

group
Character column in metadata used to group samples and applied the cutoff.

min
Numeric value indicating the minimum number of samples in each group that should have more than 0 in count matrix.

minreads
Integer minimum number of reads to consider a feature expressed.

Value

count matrix after filtering genes (features) with not enough expression in any group.

Examples

data(humanGender)
library(SummarizedExperiment)
idx <- c(1:10, 75:85)
c <- degFilter(assays(humanGender)[[1]][1:1000, idx],
  colData(humanGender)[idx,], "group", min=1)