Integrate data comming from degPattern into one data object

The simplest case is if you want to convine the pattern profile for gene expression data and proteomic data. It will use the first element as the base for the integration. Then, it will loop through clusters and run degPatterns in the second data set to detect patterns that match this one.

degMerge(matrix_list, cluster_list, metadata_list, summarize = "group",
  time = "time", col = "condition", scale = TRUE, mapping = NULL)

Arguments

matrix_list
list expression data for each element

cluster_list
list df item from degPattern output

metadata_list
list data.frames from each element with design experiment. Normally colData output

summarize
character column to use to group samples

time
character column to use as x-axes in figures

col
character column to color samples in figures

scale
boolean scale by row expression matrix

mapping
data.frame mapping table in case elements use different ID in the row.names of expression matrix. For instance, when integrating miRNA/mRNA.

Value

A data.frame with information on what genes are in each cluster in all data set, and the correlation value for each pair cluster comparison.