Make groups of genes using expression profile

Make groups of genes using expression profile

degPatterns(ma, metadata, minc = 15, summarize = "merge", time = "time",
  col = NULL, concensusCluster = TRUE, reduce = FALSE, cutoff = 0.7,
  scale = TRUE, plot = TRUE, fixy = NULL)

Arguments

ma
log2 normalized count matrix

metadata
data frame with sample information. Rownames should match ma column names row number should be the same length than p-values vector.

minc
integer minimum number of genes in a group that will be return

summarize
character column name in metadata that will be used to group replicates. If the column doesn't exist it'll merge the time and the col columns, if col doesn't exist it'll use time only. For instance, a merge between summarize and time parameters: control_point0 ... etc

time
character column name in metadata that will be used as variable that changes, normally a time variable.

col
character column name in metadata to separate samples. Normally control/mutant

concensusCluster
Indicates whether using ConsensusClusterPlus or cluster::diana()

reduce
boolean reduce number of clusters using correlation values between them.

cutoff
integer threshold for correlation expression to merge clusters (0 - 1)

scale
boolean scale the ma values by row

plot
boolean plot the clusters found

fixy
vector integers used as ylim in plot

Value

list wiht two items. df is a data.frame with two columns. The first one with genes, the second with the clusters they belong. pass_to_plot is a vector of the clusters that pass the minc cutoff.

Details

It would be used cluster::diana() function to detect a value to cut the expression based clustering at certain height or ConsensusClusterPlus. It can work with one or more groups with 2 or more several time points. The different patterns can be merged to get similar ones into only one pattern. The expression correlation of the patterns will be used to decide whether some need to be merged or not.

Examples

data(humanGender)
library(SummarizedExperiment)
ma <- assays(humanGender)[[1]][1:100,]
des <- colData(humanGender)
res <- degPatterns(ma, des, time="group", concensusCluster = FALSE)

#> Working with 100 genes.

#> Working with 100 genes after filtering: minc > 15