DEGSet

S4 class to store data from differentially expression analysis. It should be compatible with different package and stores the information in a way the methods will work with all of them.

DEGSet(resList, default)

DEGSet(resList, default)

DEGSetFromEdgeR(object, ...)

DEGSetFromDESeq2(object, ...)

# S4 method for TopTags
DEGSetFromEdgeR(object, default = "shrunken",
  extras = NULL)

# S4 method for DESeqResults
DEGSetFromDESeq2(object, default = "shrunken",
  extras = NULL)

Arguments

resList
List with results as elements containing log2FoldChange, pvalues and padj as column. Rownames should be feature names. Elements should have names.

default
The name of the element to use by default.

object
Different objects to be transformed to DEGSet.

...
Optional parameters of the generic.

extras
List of extra tables related to the same comparison.

Details

For now supporting only DESeq2::results() output. Use constructor degComps() to create the object.

The list will contain one element for each comparison done. Each element has the following structure:

DEG table

Optional table with shrunk Fold Change when it has been done.

To access the raw table use deg(dgs, "raw"), to access the shrunken table use deg(dgs, "shrunken") or just deg(dgs).

Examples

library(DESeq2)
dds <- makeExampleDESeqDataSet(betaSD = 1)
colData(dds)[["treatment"]] <- sample(colData(dds)[["condition"]], 12)
design(dds) <-  ~ condition + treatment
dds <- DESeq(dds)

#> estimating size factors

#> estimating dispersions

#> gene-wise dispersion estimates

#> mean-dispersion relationship

#> final dispersion estimates

#> fitting model and testing

res <- degComps(dds, combs = c("condition"))

#> Doing 1 element(s).

#> Doing results() for each element.

#> Doing lcfSrink() for each element.

deg(res[[1]])

#> log2 fold change (MAP): condition B vs A 
#> Wald test p-value: condition B vs A 
#> DataFrame with 1000 rows and 6 columns
#>          baseMean log2FoldChange     lfcSE      stat    pvalue      padj
#>         <numeric>      <numeric> <numeric> <numeric> <numeric> <numeric>
#> gene542     68.20         -3.479    0.3091   -11.029 2.761e-28 2.538e-25
#> gene944    398.04          2.143    0.2632     8.124 4.508e-16 2.072e-13
#> gene815    324.23          2.161    0.2991     7.221 5.146e-13 1.576e-10
#> gene927     34.19         -2.518    0.3590    -7.078 1.466e-12 3.368e-10
#> gene369    267.88          1.883    0.2767     6.800 1.045e-11 1.921e-09
#> ...           ...            ...       ...       ...       ...       ...
#> gene943     1.425        -0.2236    0.4942   -0.4366    0.6624        NA
#> gene948     1.174        -0.5389    0.4859   -1.1552    0.2480        NA
#> gene983     1.581        -0.8012    0.5152   -1.2047    0.2283        NA
#> gene995     1.403         0.3411    0.4788    0.3910    0.6958        NA
#> gene998     1.536        -0.3942    0.5076   -0.9339    0.3503        NA

deg(res[[1]], tidy = "tibble")

#> # A tibble: 1,000 x 7
#>    gene    baseMean log2FoldChange lfcSE   stat           pvalue           padj
#>  * <chr>      <dbl>          <dbl> <dbl>  <dbl>            <dbl>          <dbl>
#>  1 gene542     68.2          -3.48 0.309 -11.0          2.76e⁻²⁸       2.54e⁻²⁵
#>  2 gene944    398             2.14 0.263   8.12         4.51e⁻¹⁶       2.07e⁻¹³
#>  3 gene815    324             2.16 0.299   7.22         5.15e⁻¹³       1.58e⁻¹⁰
#>  4 gene927     34.2          -2.52 0.359 - 7.08         1.47e⁻¹²       3.37e⁻¹⁰
#>  5 gene369    268             1.88 0.277   6.80         1.05e⁻¹¹       1.92e⁻ ⁹
#>  6 gene360    126            -2.10 0.325 - 6.44         1.23e⁻¹⁰       1.88e⁻ ⁸
#>  7 gene986     71.6          -1.99 0.322 - 6.17         6.67e⁻¹⁰       8.76e⁻ ⁸
#>  8 gene866    230             1.60 0.263   6.11         1.03e⁻ ⁹       1.18e⁻ ⁷
#>  9 gene739    158             2.04 0.341   5.96         2.52e⁻ ⁹       2.57e⁻ ⁷
#> 10 gene657    253            -1.25 0.211 - 5.92         3.20e⁻ ⁹       2.94e⁻ ⁷
#> # ... with 990 more rows