DEGSet

S4 class to store data from differentially expression analysis. It should be compatible with different package and stores the information in a way the methods will work with all of them.

DEGSet(resList, default)

DEGSet(resList, default)

DEGSetFromEdgeR(object, ...)

DEGSetFromDESeq2(object, ...)

# S4 method for TopTags
DEGSetFromEdgeR(object, default = "shrunken",
  extras = NULL)

# S4 method for DESeqResults
DEGSetFromDESeq2(object, default = "shrunken",
  extras = NULL)

Arguments

resList
List with results as elements containing log2FoldChange, pvalues and padj as column. Rownames should be feature names. Elements should have names.

default
The name of the element to use by default.

object
Different objects to be transformed to DEGSet.

...
Optional parameters of the generic.

extras
List of extra tables related to the same comparison.

Details

For now supporting only DESeq2::results() output. Use constructor degComps() to create the object.

The list will contain one element for each comparison done. Each element has the following structure:

DEG table

Optional table with shrunk Fold Change when it has been done.

To access the raw table use deg(dgs, "raw"), to access the shrunken table use deg(dgs, "shrunken") or just deg(dgs).

Examples

library(DESeq2)

#> Loading required package: S4Vectors

#> Loading required package: stats4

#> Loading required package: BiocGenerics

#> Loading required package: parallel

#> 
#> Attaching package: ‘BiocGenerics’

#> The following objects are masked from ‘package:parallel’:
#> 
#>     clusterApply, clusterApplyLB, clusterCall, clusterEvalQ,
#>     clusterExport, clusterMap, parApply, parCapply, parLapply,
#>     parLapplyLB, parRapply, parSapply, parSapplyLB

#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs

#> The following objects are masked from ‘package:base’:
#> 
#>     anyDuplicated, append, as.data.frame, cbind, colMeans, colnames,
#>     colSums, do.call, duplicated, eval, evalq, Filter, Find, get, grep,
#>     grepl, intersect, is.unsorted, lapply, lengths, Map, mapply, match,
#>     mget, order, paste, pmax, pmax.int, pmin, pmin.int, Position, rank,
#>     rbind, Reduce, rowMeans, rownames, rowSums, sapply, setdiff, sort,
#>     table, tapply, union, unique, unsplit, which, which.max, which.min

#> 
#> Attaching package: ‘S4Vectors’

#> The following object is masked from ‘package:base’:
#> 
#>     expand.grid

#> Loading required package: IRanges

#> Loading required package: GenomicRanges

#> Loading required package: GenomeInfoDb

#> Loading required package: SummarizedExperiment

#> Loading required package: Biobase

#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.

#> Loading required package: DelayedArray

#> Loading required package: matrixStats

#> 
#> Attaching package: ‘matrixStats’

#> The following objects are masked from ‘package:Biobase’:
#> 
#>     anyMissing, rowMedians

#> 
#> Attaching package: ‘DelayedArray’

#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     colMaxs, colMins, colRanges, rowMaxs, rowMins, rowRanges

#> The following object is masked from ‘package:base’:
#> 
#>     apply

dds <- makeExampleDESeqDataSet(betaSD = 1)
colData(dds)[["treatment"]] <- sample(colData(dds)[["condition"]], 12)
design(dds) <-  ~ condition + treatment
dds <- DESeq(dds)

#> estimating size factors

#> estimating dispersions

#> gene-wise dispersion estimates

#> mean-dispersion relationship

#> final dispersion estimates

#> fitting model and testing

res <- degComps(dds, combs = c("condition"))

#> Doing 1 element(s).

#> Doing results() for each element.

#> Doing lcfSrink() for each element.

deg(res[[1]])

#> log2 fold change (MAP): condition B vs A 
#> Wald test p-value: condition B vs A 
#> DataFrame with 1000 rows and 6 columns
#>           baseMean log2FoldChange     lfcSE      stat    pvalue      padj
#>          <numeric>      <numeric> <numeric> <numeric> <numeric> <numeric>
#> gene202     595.74          2.693    0.3128     7.829 4.925e-15 4.058e-12
#> gene710     117.42          2.392    0.3094     6.934 4.100e-12 1.689e-09
#> gene365      62.44          2.688    0.3470     6.795 1.084e-11 2.976e-09
#> gene364     504.52          2.137    0.3042     6.564 5.238e-11 1.079e-08
#> gene92      220.78          2.058    0.3269     5.430 5.642e-08 9.299e-06
#> ...            ...            ...       ...       ...       ...       ...
#> gene984      3.561        1.18342    0.5582    1.5661    0.1173        NA
#> gene986      1.039       -0.13169    0.5500   -0.7438    0.4570        NA
#> gene989      1.841        0.79708    0.5677    1.3150    0.1885        NA
#> gene995      2.886        0.08408    0.5557    1.0914    0.2751        NA
#> gene1000     2.582       -0.53534    0.5650   -0.8062    0.4201        NA

deg(res[[1]], tidy = "tibble")

#> # A tibble: 1,000 x 7
#>    gene    baseMean log2FoldChange lfcSE  stat              pvalue         padj
#>  * <chr>      <dbl>          <dbl> <dbl> <dbl>               <dbl>        <dbl>
#>  1 gene202    596             2.69 0.313  7.83 0.00000000000000493     4.06e⁻¹²
#>  2 gene710    117             2.39 0.309  6.93 0.00000000000410        1.69e⁻ ⁹
#>  3 gene365     62.4           2.69 0.347  6.79 0.0000000000108         2.98e⁻ ⁹
#>  4 gene364    505             2.14 0.304  6.56 0.0000000000524         1.08e⁻ ⁸
#>  5 gene92     221             2.06 0.327  5.43 0.0000000564            9.30e⁻ ⁶
#>  6 gene664    110             1.97 0.350  5.20 0.000000203             2.79e⁻ ⁵
#>  7 gene577    135            -1.66 0.295 -5.17 0.000000237             2.79e⁻ ⁵
#>  8 gene564     83.3          -1.93 0.361 -5.12 0.000000308             3.10e⁻ ⁵
#>  9 gene787     71.0           1.92 0.360  5.08 0.000000376             3.10e⁻ ⁵
#> 10 gene836     49.2           2.25 0.377  5.08 0.000000371             3.10e⁻ ⁵
#> # ... with 990 more rows