DEGSet

S4 class to store data from differentially expression analysis. It should be compatible with different package and stores the information in a way the methods will work with all of them.

DEGSet(resList, default)

DEGSet(resList, default)

DEGSetFromEdgeR(object, ...)

DEGSetFromDESeq2(object, ...)

# S4 method for TopTags
DEGSetFromEdgeR(object, default = "shrunken",
  extras = NULL)

# S4 method for DESeqResults
DEGSetFromDESeq2(object, default = "shrunken",
  extras = NULL)

Arguments

resList
List with results as elements containing log2FoldChange, pvalues and padj as column. Rownames should be feature names. Elements should have names.

default
The name of the element to use by default.

object
Different objects to be transformed to DEGSet.

...
Optional parameters of the generic.

extras
List of extra tables related to the same comparison.

Details

For now supporting only DESeq2::results() output. Use constructor degComps() to create the object.

The list will contain one element for each comparison done. Each element has the following structure:

DEG table

Optional table with shrunk Fold Change when it has been done.

To access the raw table use deg(dgs, "raw"), to access the shrunken table use deg(dgs, "shrunken") or just deg(dgs).

Examples

library(DESeq2)
dds <- makeExampleDESeqDataSet(betaSD = 1)
colData(dds)[["treatment"]] <- sample(colData(dds)[["condition"]], 12)
design(dds) <-  ~ condition + treatment
dds <- DESeq(dds)

#> estimating size factors

#> estimating dispersions

#> gene-wise dispersion estimates

#> mean-dispersion relationship

#> final dispersion estimates

#> found already estimated fitted dispersions, removing these

#> fitting model and testing

res <- degComps(dds, combs = c("condition"))

#> Doing 1 element(s).

#> Doing results() for each element.

#> Doing lcfSrink() for each element.

deg(res[[1]])

#> log2 fold change (MAP): condition B vs A 
#> Wald test p-value: condition B vs A 
#> DataFrame with 1000 rows and 6 columns
#>          baseMean log2FoldChange     lfcSE      stat    pvalue      padj
#>         <numeric>      <numeric> <numeric> <numeric> <numeric> <numeric>
#> gene168    284.99          2.866    0.3258     8.945 3.716e-19 3.482e-16
#> gene248   1131.03          2.187    0.2604     8.166 3.201e-16 1.500e-13
#> gene72     135.85          2.400    0.3094     7.349 1.993e-13 6.223e-11
#> gene625     56.00         -2.350    0.3172    -7.220 5.203e-13 1.219e-10
#> gene872     52.23         -2.767    0.3721    -6.996 2.624e-12 4.918e-10
#> ...           ...            ...       ...       ...       ...       ...
#> gene941    1.2913       -0.28229    0.4818   0.08827    0.9297        NA
#> gene962    1.1453       -0.61940    0.4511  -0.60991    0.5419        NA
#> gene967    0.9839       -0.29527    0.4910  -0.28823    0.7732        NA
#> gene968    0.4834        0.11466    0.3691   0.34963    0.7266        NA
#> gene984    1.3625        0.03872    0.5366  -0.05969    0.9524        NA

deg(res[[1]], tidy = "tibble")

#> # A tibble: 1,000 x 7
#>       gene baseMean log2FoldChange  lfcSE   stat    pvalue      padj
#>  *   <chr>    <dbl>          <dbl>  <dbl>  <dbl>     <dbl>     <dbl>
#>  1 gene168   284.99          2.866 0.3258  8.945 3.716e-19 3.482e-16
#>  2 gene248  1131.03          2.187 0.2604  8.166 3.201e-16 1.500e-13
#>  3  gene72   135.85          2.400 0.3094  7.349 1.993e-13 6.223e-11
#>  4 gene625    56.00         -2.350 0.3172 -7.220 5.203e-13 1.219e-10
#>  5 gene872    52.23         -2.767 0.3721 -6.996 2.624e-12 4.918e-10
#>  6 gene851    24.73         -2.543 0.3972 -6.395 1.601e-10 2.500e-08
#>  7 gene327  5001.55          1.709 0.2707  6.277 3.455e-10 4.624e-08
#>  8 gene247    55.54         -2.316 0.3614 -6.075 1.238e-09 1.450e-07
#>  9  gene70   878.25          1.263 0.2037  6.032 1.621e-09 1.687e-07
#> 10 gene514   120.11          1.854 0.2912  5.979 2.241e-09 2.099e-07
#> # ... with 990 more rows