DEGSet class

S4 class to store data from differentially expression analysis. It should be compatible with different package and stores the information in a way the methods will work with all of them.

Details

For now supporting only DESeq2::results() output. Use constructor degComps() to create the object.

The list will contain one element for each comparison done. Each element has the following structure:

DEG table

Optional table with shrunk Fold Change when it has been done.

To access the raw table use deg(dgs, "raw")``, to access the shrunken table usedeg(dgs, "shrunk")`.

Examples

library(DESeq2)
dds <- makeExampleDESeqDataSet(betaSD = 1)
colData(dds)[["treatment"]] <- sample(colData(dds)[["condition"]], 12)
design(dds) <-  ~ condition + treatment
dds <- DESeq(dds)

#> estimating size factors

#> estimating dispersions

#> gene-wise dispersion estimates

#> mean-dispersion relationship

#> final dispersion estimates

#> fitting model and testing

res <- degComps(dds, combs = c("condition"))

#> Doing 1 element(s).

#> Doing results() for each element.

#> Doing lcfSrink() for each element.

deg(res[[1]])

#> log2 fold change (MAP): condition B vs A 
#> Wald test p-value: condition B vs A 
#> DataFrame with 1000 rows and 5 columns
#>          baseMean log2FoldChange      stat    pvalue      padj
#>         <numeric>      <numeric> <numeric> <numeric> <numeric>
#> gene254   1297.34       -0.07239     8.505 1.823e-17 1.741e-14
#> gene174    245.74       -0.21299     7.912 2.542e-15 1.214e-12
#> gene878     51.98       -0.39723    -7.671 1.702e-14 5.417e-12
#> gene631     55.75        0.09504    -7.310 2.678e-13 6.393e-11
#> gene253     54.27       -0.40900    -6.885 5.784e-12 1.105e-09
#> ...           ...            ...       ...       ...       ...
#> gene843    0.3203         1.1478   0.36759    0.7132        NA
#> gene846    0.6453         1.1056   0.09192    0.9268        NA
#> gene863    0.3218         0.5457  -0.34644    0.7290        NA
#> gene928    0.4599         0.1447   0.59410    0.5524        NA
#> gene974    0.4876        -1.6675   0.35872    0.7198        NA

deg(res[[1]], tidy = "tibble")

#> # A tibble: 1,000 x 6
#>       gene baseMean log2FoldChange   stat    pvalue      padj
#>  *   <chr>    <dbl>          <dbl>  <dbl>     <dbl>     <dbl>
#>  1 gene254  1297.34       -0.07239  8.505 1.823e-17 1.741e-14
#>  2 gene174   245.74       -0.21299  7.912 2.542e-15 1.214e-12
#>  3 gene878    51.98       -0.39723 -7.671 1.702e-14 5.417e-12
#>  4 gene631    55.75        0.09504 -7.310 2.678e-13 6.393e-11
#>  5 gene253    54.27       -0.40900 -6.885 5.784e-12 1.105e-09
#>  6  gene78   154.21       -0.18257  6.798 1.059e-11 1.686e-09
#>  7 gene796    70.54       -1.35350  6.625 3.471e-11 4.735e-09
#>  8 gene114   107.95       -0.50591 -6.361 1.999e-10 2.386e-08
#>  9  gene80   667.61       -1.04634  6.249 4.121e-10 4.373e-08
#> 10 gene529    32.77       -1.48273 -6.072 1.261e-09 1.205e-07
#> # ... with 990 more rows