Automatize the use of results() for multiple comparisons

This function will extract the output of DESeq2::results() and DESeq2::lfcShrink() for multiple comparison using:

degComps(dds, combs = NULL, contrast = NULL, alpha = 0.05, skip = FALSE,
  type = "normal", pairs = FALSE)

Arguments

dds
DESeq2::DESeqDataSet obcject.

combs
Optional vector indicating the coefficients or columns fom colData(dds) to create group comparisons.

contrast
Optional vector to specify contrast. See DESeq2::results().

alpha
Numeric value used in independent filtering in DESeq2::results().

skip
Boolean to indicate whether skip shrinkage. For instance when it comes from LRT method.

type
Type of shrinkage estimator. See DESeq2::results().

pairs
Boolean to indicate whether create all comparisons or only use the coefficient already created from DESeq2::resultsNames().

Value

DEGSet with unSrunken and Srunken results.

Details

coefficients

contrast

Multiple columns in colData that match coefficients

Multiple columns in colData to create all possible contrasts

Examples

library(DESeq2)
dds <- makeExampleDESeqDataSet(betaSD=1)
colData(dds)[["treatment"]] <- sample(colData(dds)[["condition"]], 12)
  design(dds) <-  ~ condition + treatment
dds <- DESeq(dds)

#> estimating size factors

#> estimating dispersions

#> gene-wise dispersion estimates

#> mean-dispersion relationship

#> final dispersion estimates

#> fitting model and testing

res <- degComps(dds, combs = c("condition", 2),
              contrast = list("treatment_B_vs_A", c("condition", "A", "B")))

#> Doing 3 element(s).

#> Doing results() for each element.

#> Doing lcfSrink() for each element.