Method to get the significant genes

Function to get the features that are significant according to some thresholds from a DEGSet, DESeq2::DESeqResults and edgeR::topTags.

significants(object, ...)

# S4 method for DEGSet
significants(object, padj = 0.05, fc = 0,
  direction = NULL, full = FALSE, ...)

# S4 method for DESeqResults
significants(object, padj = 0.05, fc = 0,
  direction = NULL, full = FALSE, ...)

# S4 method for TopTags
significants(object, padj = 0.05, fc = 0,
  direction = NULL, full = FALSE, ...)

# S4 method for list
significants(object, padj = 0.05, fc = 0,
  direction = NULL, full = FALSE, newFDR = FALSE, ...)

Arguments

object
DEGSet

...
Passed to deg. Default: value = NULL. Value can be 'raw', 'shrunken'.

padj
Cutoff for the FDR column.

fc
Cutoff for the log2FC column.

direction
Whether to take down/up/ignore. Valid arguments are down, up and NULL.

full
Whether to return full table or not.

newFDR
Whether to recalculate the FDR or not. See https://support.bioconductor.org/p/104059/#104072. Only used when a list is giving to the method.

Value

a dplyr::tbl_df data frame. gene column has the feature name. In the case of using this method with the results from degComps, log2FoldChange has the higher foldChange from the comparisons, and padj has the padj associated to the previous column. Then, there is two columns for each comparison, one for the log2FoldChange and another for the padj.

Examples

library(DESeq2)
library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following object is masked from ‘package:matrixStats’:
#> 
#>     count

#> The following object is masked from ‘package:Biobase’:
#> 
#>     combine

#> The following objects are masked from ‘package:GenomicRanges’:
#> 
#>     intersect, setdiff, union

#> The following object is masked from ‘package:GenomeInfoDb’:
#> 
#>     intersect

#> The following objects are masked from ‘package:IRanges’:
#> 
#>     collapse, desc, intersect, setdiff, slice, union

#> The following objects are masked from ‘package:S4Vectors’:
#> 
#>     first, intersect, rename, setdiff, setequal, union

#> The following objects are masked from ‘package:BiocGenerics’:
#> 
#>     combine, intersect, setdiff, union

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

dds <- makeExampleDESeqDataSet(betaSD=1)
colData(dds)[["treatment"]] <- sample(colData(dds)[["condition"]], 12)
  design(dds) <-  ~ condition + treatment
dds <- DESeq(dds)

#> estimating size factors

#> estimating dispersions

#> gene-wise dispersion estimates

#> mean-dispersion relationship

#> final dispersion estimates

#> fitting model and testing

res <- degComps(dds, contrast = list("treatment_B_vs_A",
                                     c("condition", "A", "B")))

#> Doing 2 element(s).

#> Doing results() for each element.

#> Doing lcfSrink() for each element.

significants(res, full = TRUE) %>% head

#> # A tibble: 6 x 5
#>   gene    log2FoldChange    padj log2FoldChange_condition_… padj_condition_A_v…
#>   <chr>            <dbl>   <dbl>                      <dbl>               <dbl>
#> 1 gene100          -1.35 0.00103                      -1.35             0.00103
#> 2 gene10…          -1.40 0.00384                      -1.40             0.00384
#> 3 gene104          -1.49 0.00431                      -1.49             0.00431
#> 4 gene11           -2.54 0.00170                      -2.54             0.00170
#> 5 gene112           1.39 0.0122                        1.39             0.0122 
#> 6 gene114          -1.22 0.0100                       -1.22             0.0100 

significants(res, full = TRUE, padj = 1) %>% head # all genes

#> # A tibble: 6 x 7
#>   gene  log2FoldChange  padj log2FoldChange_… log2FoldChange_… padj_condition_…
#>   <chr>          <dbl> <dbl>            <dbl>            <dbl>            <dbl>
#> 1 gene1         1.14   0.153            1.14           -0.171           0.153  
#> 2 gene…         0.0473 0.996           NA               0.0473         NA      
#> 3 gene…         0.325  0.976           -1.35            0.325           0.00103
#> 4 gene…         0.419  0.931           -1.40            0.419           0.00384
#> 5 gene…        -0.185  0.996           -0.343          -0.185           0.706  
#> 6 gene…         0.641  0.479            0.641          -0.192           0.479  
#> # ... with 1 more variable: padj_treatment_B_vs_A <dbl>