Method to get the significant genes

Function to get the features that are significant according to some thresholds from a DEGSet, DESeq2::DESeqResults and edgeR::topTags.

significants(object, ...)

# S4 method for DEGSet
significants(object, padj = 0.05, fc = 0,
  direction = NULL, full = FALSE, ...)

# S4 method for DESeqResults
significants(object, padj = 0.05, fc = 0,
  direction = NULL, full = FALSE, ...)

# S4 method for TopTags
significants(object, padj = 0.05, fc = 0,
  direction = NULL, full = FALSE, ...)

# S4 method for list
significants(object, padj = 0.05, fc = 0,
  direction = NULL, full = FALSE, newFDR = FALSE, ...)

Arguments

object
DEGSet

...
Passed to deg. Default: value = NULL. Value can be 'raw', 'shrunken'.

padj
Cutoff for the FDR column.

fc
Cutoff for the log2FC column.

direction
Whether to take down/up/ignore. Valid arguments are down, up and NULL.

full
Whether to return full table or not.

newFDR
Whether to recalculate the FDR or not. See https://support.bioconductor.org/p/104059/#104072. Only used when a list is giving to the method.

Examples

library(DESeq2)
library(dplyr)
dds <- makeExampleDESeqDataSet(betaSD=1)
colData(dds)[["treatment"]] <- sample(colData(dds)[["condition"]], 12)
  design(dds) <-  ~ condition + treatment
dds <- DESeq(dds)

#> estimating size factors

#> estimating dispersions

#> gene-wise dispersion estimates

#> mean-dispersion relationship

#> final dispersion estimates

#> fitting model and testing

res <- degComps(dds, contrast = list("treatment_B_vs_A",
                                     c("condition", "A", "B")))

#> Doing 2 element(s).

#> Doing results() for each element.

#> Doing lcfSrink() for each element.

significants(res, full = TRUE) %>% head

#> # A tibble: 6 x 3
#>   gene    log2FoldChange_condition_A_vs_B padj_condition_A_vs_B
#>   <chr>                             <dbl>                 <dbl>
#> 1 gene10                           -1.35            0.0477     
#> 2 gene109                          -2.03            0.000000188
#> 3 gene111                          -1.45            0.0301     
#> 4 gene116                          -1.37            0.00602    
#> 5 gene12                           -1.60            0.0177     
#> 6 gene121                           0.690           0.0410     

significants(res, full = TRUE, padj = 1) %>% head # all genes

#> # A tibble: 6 x 5
#>   gene     log2FoldChange_c… log2FoldChange_… padj_condition_… padj_treatment_…
#>   <chr>                <dbl>            <dbl>            <dbl>            <dbl>
#> 1 gene1              - 0.282          -0.734            0.939             0.986
#> 2 gene10             - 1.35           -0.650            0.0477            1.000
#> 3 gene100             NA               0.129           NA                 1.000
#> 4 gene1000            NA               0.101           NA                 1.000
#> 5 gene101            - 0.127          -0.0294           0.924             1.000
#> 6 gene102            - 0.512           0.0662           0.529             1.000