Method to get the significant genes

Function to get the features that are significant according to some thresholds from a DEGSet, DESeq2::DESeqResults and edgeR::topTags.

significants(object, ...)

# S4 method for DEGSet
significants(object, padj = 0.05, fc = 0,
  direction = NULL, full = FALSE, ...)

# S4 method for DESeqResults
significants(object, padj = 0.05, fc = 0,
  direction = NULL, full = FALSE, ...)

# S4 method for TopTags
significants(object, padj = 0.05, fc = 0,
  direction = NULL, full = FALSE, ...)

# S4 method for list
significants(object, padj = 0.05, fc = 0,
  direction = NULL, full = FALSE, newFDR = FALSE, ...)

Arguments

object
DEGSet

...
Passed to deg. Default: value = NULL. Value can be 'raw', 'shrunken'.

padj
Cutoff for the FDR column.

fc
Cutoff for the log2FC column.

direction
Whether to take down/up/ignore. Valid arguments are down, up and NULL.

full
Whether to return full table or not.

newFDR
Whether to recalculate the FDR or not. See https://support.bioconductor.org/p/104059/#104072. Only used when a list is giving to the method.

Examples

library(DESeq2)
library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following object is masked from ‘package:matrixStats’:
#> 
#>     count

#> The following objects are masked from ‘package:Biobase’:
#> 
#>     combine, exprs

#> The following objects are masked from ‘package:GenomicRanges’:
#> 
#>     intersect, setdiff, union

#> The following object is masked from ‘package:GenomeInfoDb’:
#> 
#>     intersect

#> The following objects are masked from ‘package:IRanges’:
#> 
#>     collapse, desc, intersect, setdiff, slice, union

#> The following objects are masked from ‘package:S4Vectors’:
#> 
#>     first, intersect, rename, setdiff, setequal, union

#> The following objects are masked from ‘package:BiocGenerics’:
#> 
#>     combine, intersect, setdiff, union

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

dds <- makeExampleDESeqDataSet(betaSD=1)
colData(dds)[["treatment"]] <- sample(colData(dds)[["condition"]], 12)
  design(dds) <-  ~ condition + treatment
dds <- DESeq(dds)

#> estimating size factors

#> estimating dispersions

#> gene-wise dispersion estimates

#> mean-dispersion relationship

#> final dispersion estimates

#> fitting model and testing

res <- degComps(dds, contrast = list("treatment_B_vs_A",
                                     c("condition", "A", "B")))

#> Doing 2 element(s).

#> Doing results() for each element.

#> Doing lcfSrink() for each element.

significants(res, full = TRUE) %>% head

#> # A tibble: 6 x 3
#>   gene    log2FoldChange_condition_A_vs_B padj_condition_A_vs_B
#>   <chr>                             <dbl>                 <dbl>
#> 1 gene102                            2.67         0.00942      
#> 2 gene109                           -2.06         0.00801      
#> 3 gene110                           -1.15         0.0113       
#> 4 gene119                           -1.24         0.00183      
#> 5 gene121                            1.27         0.0114       
#> 6 gene127                           -2.01         0.00000000868

significants(res, full = TRUE, padj = 1) %>% head # all genes

#> # A tibble: 6 x 5
#>   gene     log2FoldChange_c… log2FoldChange_… padj_condition_… padj_treatment_…
#>   <chr>                <dbl>            <dbl>            <dbl>            <dbl>
#> 1 gene1              -0.802            -0.142          0.240              0.992
#> 2 gene10              0.509            -0.111          0.536              0.992
#> 3 gene100             0.0922            0.257          0.995              0.935
#> 4 gene1000           -0.175            -0.526          0.654              0.627
#> 5 gene101             0.911             0.551          0.0567             0.670
#> 6 gene102             2.67             -0.255          0.00942            0.992