Method to get the significant genes

Function to get the features that are significant according to some thresholds from a DEGSet, DESeq2::DESeqResults and edgeR::topTags.

significants(object, ...)

# S4 method for DEGSet
significants(object, padj = 0.05, fc = 0,
  direction = NULL, full = FALSE, ...)

# S4 method for DESeqResults
significants(object, padj = 0.05, fc = 0,
  direction = NULL, full = FALSE, ...)

# S4 method for TopTags
significants(object, padj = 0.05, fc = 0,
  direction = NULL, full = FALSE, ...)

# S4 method for list
significants(object, padj = 0.05, fc = 0,
  direction = NULL, full = FALSE, newFDR = FALSE, ...)

Arguments

object
DEGSet

...
Passed to deg. Default: value = NULL. Value can be 'raw', 'shrunken'.

padj
Cutoff for the FDR column.

fc
Cutoff for the log2FC column.

direction
Whether to take down/up/ignore. Valid arguments are down, up and NULL.

full
Whether to return full table or not.

newFDR
Whether to recalculate the FDR or not. See https://support.bioconductor.org/p/104059/#104072. Only used when a list is giving to the method.

Examples

library(DESeq2)
library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following object is masked from ‘package:testthat’:
#> 
#>     matches

#> The following object is masked from ‘package:matrixStats’:
#> 
#>     count

#> The following objects are masked from ‘package:Biobase’:
#> 
#>     combine, exprs

#> The following objects are masked from ‘package:GenomicRanges’:
#> 
#>     intersect, setdiff, union

#> The following object is masked from ‘package:GenomeInfoDb’:
#> 
#>     intersect

#> The following objects are masked from ‘package:IRanges’:
#> 
#>     collapse, desc, intersect, setdiff, slice, union

#> The following objects are masked from ‘package:S4Vectors’:
#> 
#>     first, intersect, rename, setdiff, setequal, union

#> The following objects are masked from ‘package:BiocGenerics’:
#> 
#>     combine, intersect, setdiff, union

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

dds <- makeExampleDESeqDataSet(betaSD=1)
colData(dds)[["treatment"]] <- sample(colData(dds)[["condition"]], 12)
  design(dds) <-  ~ condition + treatment
dds <- DESeq(dds)

#> estimating size factors

#> estimating dispersions

#> gene-wise dispersion estimates

#> mean-dispersion relationship

#> final dispersion estimates

#> found already estimated fitted dispersions, removing these

#> fitting model and testing

res <- degComps(dds, contrast = list("treatment_B_vs_A",
                                     c("condition", "A", "B")))

#> Doing 2 element(s).

#> Doing results() for each element.

#> Doing lcfSrink() for each element.

significants(res, full = TRUE) %>% head

#> # A tibble: 6 x 3
#>       gene log2FoldChange_condition.A.vs.B padj_condition.A.vs.B
#>      <chr>                           <dbl>                 <dbl>
#> 1   gene10                          -1.584              0.001076
#> 2 gene1000                          -1.756              0.003003
#> 3  gene102                           1.659              0.010452
#> 4  gene103                           2.053              0.003003
#> 5  gene105                          -1.600              0.005062
#> 6  gene110                          -1.189              0.023600

significants(res, full = TRUE, padj = 1) %>% head # all genes

#> # A tibble: 6 x 5
#>       gene log2FoldChange_condition.A.vs.B log2FoldChange_treatment.B.vs.A
#>      <chr>                           <dbl>                           <dbl>
#> 1    gene1                        -0.33669                        -0.06179
#> 2   gene10                        -1.58413                        -0.10493
#> 3  gene100                         0.09264                         0.07912
#> 4 gene1000                        -1.75551                         0.39559
#> 5  gene101                         0.60857                        -0.13184
#> 6  gene102                         1.65943                        -0.44498
#> # ... with 2 more variables: padj_condition.A.vs.B <dbl>,
#> #   padj_treatment.B.vs.A <dbl>